+++ src/test/java/FactorialTest.java
@@ -6,10 +6,10 @@ public class FactorialTest {
     @Test
     public void testFactorialFive(){
         // given
-        Integer expected = 120;
+        long expected = 120;
 
         //when
-        Integer actual = Factorial.factorial(5);
+        long actual = Factorial.factorial(5);
 
         //then
         Assert.assertEquals(expected, actual);
@@ -18,10 +18,10 @@ public class FactorialTest {
     @Test
     public void testFactorialZero(){
         //given
-        Integer expected = 1;
+        long expected = 1;
 
         //when
-        Integer actual = Factorial.factorial(0);
+        long actual = Factorial.factorial(0);
 
         //then
         Assert.assertEquals(expected, actual);
